finished some detail on the Links
